// src/parser/raw-tree.d.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface CTree {
  title?: string
  children?: (Token | CTree)[]
  [k: string]: unknown
}
export interface Token {
  lexeme: string
  row: number
  col: number
  tokenClass: string
  parent: null
  children: null
  keyword?: boolean
  [k: string]: unknown
}